requirements.txt
torch
torchvision
torchaudio
--extra-index-url https://download.pytorch.org/whl/cu121
ipywidgets
jupyter
ftfy
regex
tqdm
matplotlib
opencv-python
scipy
scikit-image
numpy<2.0
pillow
git+https://github.com/openai/CLIP.git
transformers
einops
timm 